Update run.csx so it will work in modern Azure
Note: Will still have to change your Azure Function to runtime ~1.
Note: Made changes so that this will work on any generic URL, rather than OneDrive (I do not know if this will work on OneDrive, but I figured this would be more useful to most people.)
Made changes so that my Azure function would compile and return the expected data. 
This GitHub looks like its no longer being maintained but if anyone else needs this functionality, this *should* work for you, as long as you change the runtime back and you correctly specify the arguments in FunctionConfig.json.